move binarySearch to SearchUtil
